Refactor DynamoDB utilities into separate modules
Reorganized and modularized DynamoDB utility functions. Moved the functions from `dynamo-db-util.ts` into dedicated modules for better structure and maintainability. Also added type definitions in a new file `dynamo-db-types.ts` and updated the seed generator to temporarily disable binary set generation due to duplication issues.
