+++ src/dynamo-db/dynamo-db-types.ts
@@ -1,0 +1,5 @@
+import type { TableDescription } from '@aws-sdk/client-dynamodb/dist-types';
+
+export type TDynamoDBSchema = TableDescription & {
+    partitionKey: string;
+};
